test(cli): split app builder style tests
