+++ packages/core/test/cli/app-builder/styles/dev.ts
@@ -1,0 +1,97 @@
+import {
+  describe,
+  test,
+  expect,
+  spawn,
+  mkdtempSync,
+  writeFileSync,
+  rmSync,
+  mkdirSync,
+  tmpdir,
+  join,
+  CLI,
+  RUNTIME,
+  waitForServer,
+  findFreePort,
+} from '../helpers';
+
+describe('CLI: Vite-style app builder > styles > dev', () => {
+  test('dev applies PostCSS config and serves transformed CSS', async () => {
+    const dir = mkdtempSync(join(tmpdir(), 'zntc-app-dev-postcss-'));
+    mkdirSync(join(dir, 'src'), { recursive: true });
+    writeFileSync(
+      join(dir, 'index.html'),
+      '<title>dev</title><link rel="stylesheet" href="/src/style.css"><script type="module" src="/src/main.ts"></script>',
+    );
+    writeFileSync(join(dir, 'src', 'main.ts'), 'console.log("ok");');
+    writeFileSync(join(dir, 'src', 'style.css'), '.x{color:red}');
+    writeFileSync(
+      join(dir, 'postcss.config.mjs'),
+      [
+        'export default {',
+        '  plugins: [',
+        "    { postcssPlugin: 'zntc-dev-postcss', Once(root) { root.append({ selector: '.dev-postcss-ok', nodes: [] }); } },",
+        '  ],',
+        '};',
+      ].join('\n'),
+    );
+
+    const port = await findFreePort();
+    const proc = spawn(RUNTIME, [CLI, 'dev', dir, `--port=${port}`], { cwd: dir });
+    const stderrChunks: string[] = [];
+    proc.stderr?.on('data', (chunk) => stderrChunks.push(chunk.toString()));
+    await waitForServer(port);
+    try {
+      const html = await fetch(`http://localhost:${port}/`).then((r) => r.text());
+      expect(html).toContain('<title>dev</title>');
+      expect(html).toContain('/__zntc_app_dev_hmr__');
+      expect(html).toContain('href="/src/style.css"');
+      const css = await fetch(`http://localhost:${port}/src/style.css`).then((r) => r.text());
+      expect(css).toContain('.dev-postcss-ok');
+      const stderrText = stderrChunks.join('');
+      expect(stderrText).toContain('[postcss] processed 1 CSS file');
+      expect(stderrText).not.toContain('skipped in dev mode');
+    } finally {
+      proc.kill();
+      rmSync(dir, { recursive: true, force: true });
+    }
+  });
+
+  test('dev CSS source edit emits css-update instead of full-reload', async () => {
+    const dir = mkdtempSync(join(tmpdir(), 'zntc-app-dev-css-hmr-'));
+    mkdirSync(join(dir, 'src'), { recursive: true });
+    writeFileSync(
+      join(dir, 'index.html'),
+      '<link rel="stylesheet" href="/src/style.css"><script type="module" src="/src/main.ts"></script>',
+    );
+    writeFileSync(join(dir, 'src', 'main.ts'), 'console.log("ok");');
+    writeFileSync(join(dir, 'src', 'style.css'), '.x{color:red}');
+    writeFileSync(join(dir, 'postcss.config.mjs'), 'export default { plugins: [] };\n');
+
+    const port = await findFreePort();
+    const proc = spawn(RUNTIME, [CLI, 'dev', dir, `--port=${port}`], { cwd: dir });
+    await waitForServer(port);
+    try {
+      const messagePromise = new Promise<any>((resolve) => {
+        const ws = new WebSocket(`ws://localhost:${port}/__hmr`);
+        ws.onmessage = (event) => {
+          const msg = JSON.parse(String(event.data));
+          if (msg.type === 'css-update' || msg.type === 'full-reload') {
+            ws.close();
+            resolve(msg);
+          }
+        };
+        ws.onerror = () => resolve({ type: 'error' });
+        setTimeout(() => resolve({ type: 'timeout' }), 10000);
+      });
+      await new Promise((r) => setTimeout(r, 300));
+      writeFileSync(join(dir, 'src', 'style.css'), '.x{color:blue}');
+      const msg = await messagePromise;
+      expect(msg.type).toBe('css-update');
+      expect(msg.href).toBe('/src/style.css');
+    } finally {
+      proc.kill();
+      rmSync(dir, { recursive: true, force: true });
+    }
+  });
+});
